Renamed `.js` to `.ts`
