+++ .gitignore
@@ -28,3 +28,5 @@ node_modules
 .lock-wscript
 
 test-page-webpack/dist
+
+.idea/
