+++ src/record/anonymous-record.ts
